Wrong return type for sparsebundlefs_getsize.

+++ src/sparsebundlefs/sparsebundlefs.h
@@ -13,7 +13,7 @@ extern "C" {
 int sparsebundlefs_getdatasize();
 int sparsebundlefs_open(const char* path, const char*password, void* data);
 size_t sparsebundlefs_getblocksize(void* sparsebundle_data);
-size_t sparsebundlefs_getsize(void* sparsebundle_data);
+off_t sparsebundlefs_getsize(void* sparsebundle_data);
 char* sparsebundlefs_getpath(void* sparsebundle_data);
 size_t sparsebundlefs_read(void* sparsebundle_data, uint8_t *buffer, size_t nbytes, off_t offset);
 int sparsebundlefs_close(void* sparsebundle_data);
